refactor: request 경로 이동
