+++ src/test/java/kitchenpos/fixture/MenuProductFixtures.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.dto.MenuProductCreateRequest;
+import kitchenpos.dto.request.MenuProductCreateRequest;
 
 public class MenuProductFixtures {
